requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
django-filter==25.2
django-jazzmin==3.0.4
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
packaging==26.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
django-jazzmin==2.6.0 
